[OpenViBE] Fix some compilation warnings in build of 64 bits. * Replace some unsigned int with "size_t" in several local method * Change signature of CString length method to return a size_t
Tracker reference: CORE-1270

+++ common/include/ovCResampler.h
@@ -333,22 +333,22 @@ namespace Common
 			 */
 			size_t resample_channel_wise(const ICallback& rCallback, const TFloat* pInputSample, size_t ui32InputSampleCount)
 			{
-				size_t l_iCount;
+				int l_iCount;
 				bool l_bIsFirstChannel = true;
 
 				std::vector < double > l_vInputBuffer(ui32InputSampleCount);
 				std::vector < TFloat > l_vOutputBuffer;
 
-				for(size_t j = 0; j < m_ui32ChannelCount; j++)
+				for (int j = 0; j < m_ui32ChannelCount; j++)
 				{
-					for(size_t k = 0; k < ui32InputSampleCount; k++)
+					for (int k = 0; k < ui32InputSampleCount; k++)
 					{
 						l_vInputBuffer[k] = static_cast<double>(pInputSample[k*m_ui32ChannelCount + j]);
 					}
 
 					l_iCount = 0;
 					double* l_pResamplerOutputBuffer;
-					l_iCount = m_vResampler[j] -> process(&l_vInputBuffer[0], ui32InputSampleCount, l_pResamplerOutputBuffer);
+					l_iCount = m_vResampler[j] -> process(&l_vInputBuffer[0], static_cast<int>(ui32InputSampleCount), l_pResamplerOutputBuffer);
 
 					if(l_bIsFirstChannel)
 					{
@@ -356,13 +356,13 @@ namespace Common
 						l_bIsFirstChannel = false;
 					}
 
-					for(size_t k = 0; k < l_iCount; k++)
+					for (int k = 0; k < l_iCount; k++)
 					{
 						l_vOutputBuffer[k*m_ui32ChannelCount + j] = static_cast<TFloat>(l_pResamplerOutputBuffer[k]);
 					}
 				}
 				
-				for(size_t k = 0; k < l_iCount; k++)
+				for (int k = 0; k < l_iCount; k++)
 				{
 					rCallback.processResampler(&l_vOutputBuffer[k*m_ui32ChannelCount], m_ui32ChannelCount);
 				}
@@ -396,7 +396,7 @@ namespace Common
 
 					l_iCount = 0;
 					double* l_pResamplerOutputBuffer;
-					l_iCount = m_vResampler[j] -> process(&l_vInputBuffer[0], ui32InputSampleCount, l_pResamplerOutputBuffer);
+					l_iCount = m_vResampler[j]->process(&l_vInputBuffer[0], static_cast<int>(ui32InputSampleCount), l_pResamplerOutputBuffer);
 
 					if(l_bIsFirstChannel)
 					{
